requirements.txt
Flask==2.2.5
Flask-Caching==2.1.0
Flask-Session==0.5.0
gunicorn==21.2.0
Jinja2==3.1.2
mysql-connector-python==8.2.0
mysql_connector_repackaged==0.3.1
networkx==3.0
numpy==1.25.2
redis==4.4.2
